+++ src/chart/visualizations/SankeyVisualization.tsx
@@ -5,6 +5,7 @@ import {evaluateRulesOnDict, evaluateRulesOnNode} from '../../report/ReportRuleE
 import {categoricalColorSchemes} from "../../config/ColorConfig";
 import {valueIsArray, valueIsNode, valueIsPath, valueIsRelationship} from "../../report/ReportRecordProcessing";
 import NeoCodeViewerComponent from '../../component/editor/CodeViewerComponent';
+import {isCyclic} from '../visualizations/Utils';
 
 export default function SankeyVisualization(props: ExtendedChartReportProps) {
     const { records, first } = props
@@ -156,52 +157,6 @@ export default function SankeyVisualization(props: ExtendedChartReportProps) {
         return <NeoCodeViewerComponent value={"No relationship weights found. \nDefine a numeric 'Relationship Property' in the \nreport's advanced settings to view the sankey diagram."}></NeoCodeViewerComponent>
     }
 
-    function isCyclicUtil(i,visited,recStack, adj)
-    {
-        // Mark the current node as visited and
-        // part of recursion stack
-        if (recStack.get(i))
-            return true;
-
-        if (visited.get(i))
-            return false;
-
-        visited.set(i, true);
-        recStack.set(i, true);
-
-        let childrens = adj.get(i);
-
-        for (const children in childrens)
-            if (isCyclicUtil(childrens[children], visited, recStack, adj))
-                return true;
-
-        recStack.set(i, false);
-
-        return false;
-    }
-
-
-    const isCyclic = (graph) => {
-        let visited = new Map();
-        let recStack = new Map();
-        let adj = new Map();
-
-        graph.nodes.forEach( (node) => {
-            visited.set(node["id"], false);
-            recStack.set(node["id"], false);
-            adj.set(node["id"], []);
-        })
-
-        graph.links.forEach( (link) => {
-            adj.get(link["source"]).push(link["target"]);
-        })
-
-        for (const idx in graph.nodes)
-            if (isCyclicUtil(graph.nodes[idx]["id"], visited, recStack, adj))
-                return true;
-
-        return false;
-    }
 
     if(data && data.nodes && data.links && isCyclic(data)){
         return <NeoCodeViewerComponent value={"Please be careful with the data you use for this chart as it does not support cyclic dependencies."}></NeoCodeViewerComponent>;
